More Test Cleanups (#1940)
* git status

* Move property tests into test package

* Move HalBuilderTest into graql which contains the HAl functionality

* Move ProcessWideLockProviderTest into grakn-engine

* Get rid of test which was secretly a duplicate

* Rip apart graph test
