+++ grakn-graql/src/test/java/ai/grakn/graql/internal/hal/HALBuilderTest.java
@@ -1,4 +1,22 @@
-package ai.grakn.test.graql.hal;
+/*
+ * Grakn - A Distributed Semantic Database
+ * Copyright (C) 2016  Grakn Labs Limited
+ *
+ * Grakn is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * Grakn is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with Grakn. If not, see <http://www.gnu.org/licenses/gpl.txt>.
+ */
+
+package ai.grakn.graql.internal.hal;
 
 import ai.grakn.GraknGraph;
 import ai.grakn.concept.Concept;
@@ -61,7 +79,6 @@ public class HALBuilderTest {
             }
         });
     }
-
 
     @Test
     public void whenUseSelectInQueryUsingInference_EnsureWeReceiveAValidHALResponse() {
