+++ grakn-test/src/test/java/ai/grakn/test/property/ResourceTypePropertyTest.java
@@ -14,10 +14,9 @@
  *
  * You should have received a copy of the GNU General Public License
  * along with Grakn. If not, see <http://www.gnu.org/licenses/gpl.txt>.
- *
  */
 
-package ai.grakn.property;
+package ai.grakn.test.property;
 
 import ai.grakn.concept.ResourceType;
 import ai.grakn.exception.GraphOperationException;
